Finishing touches
Enum Script exists
Cleanup in CardList and Player
