+++ BangerProjekt/Assets/Scripts/GeneralScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 801cc5b87ecf651d2b1b512cc3b64d12
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
