+++ BangerProjekt/Assets/Scripts/GeneralScripts/Enums.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20142ef8e4941c6b2ad36fae9e7184ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
